+++ CHANGES.md
@@ -5,6 +5,7 @@
 **BugFix**
 
 - Fixes a bug where using solid subsets when launching pipeline runs would fail config validation.
+- (dagster-gcp) allow multiple "bq_solid_for_queries" solids to co-exist in a pipeline
 
 ## 0.7.12